% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BED_To_Dosage.R
\name{Bed_To_dosage}
\alias{Bed_To_dosage}
\title{Bed_To_dosage function to convert previously generated H5 files into plink files}
\usage{
Bed_To_dosage(
  input.dir = ".",
  output.dir = NULL,
  prefix = NULL,
  poi.chunk.size = 1000
)
}
\arguments{
\item{input.dir}{folder location where input files reside}

\item{output.dir}{folder location where plink files will be saved}

\item{prefix}{a character denoting which set of h5 files to convert (first set is used by default)}

\item{poi.chunk.size}{integer denoting chunk.size to be used during conversion}
}
\value{
integer 1 when successful
}
\description{
Bed_To_dosage function to convert previously generated H5 files into plink files
}
