% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_To_BED_Conversion_Function.R
\name{convert_H5_To_Bed}
\alias{convert_H5_To_Bed}
\title{convert_H5_To_Bed function to convert previously generated H5 files into plink files}
\usage{
convert_H5_To_Bed(
  input.dir = ".",
  output.dir = NULL,
  prefix = NULL,
  poi.chunk.size = 1000,
  seed = 12133,
  num.poi = NULL,
  num.ind = NULL
)
}
\arguments{
\item{input.dir}{folder location where input files reside}

\item{output.dir}{folder location where plink files will be saved}

\item{prefix}{a character denoting which set of h5 files to convert (first set is used by default)}

\item{poi.chunk.size}{integer denoting chunk.size to be used during conversion}

\item{seed}{random number generator seed}

\item{num.poi}{number of poi used during conversion (default=NULL, equivalent to all)}

\item{num.ind}{number of ind used during conversion (default=NULL, equivalent to all)}
}
\value{
integer 1 when successful
}
\description{
convert_H5_To_Bed function to convert previously generated H5 files into plink files
}
