% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastR_hdf5convert.R
\name{fastR_hdf5convert}
\alias{fastR_hdf5convert}
\title{hdf5convert a function to convert textual data to hdf5 format supported by FastReg()}
\usage{
fastR_hdf5convert(
  dataFile,
  h5File,
  headerRow = 1,
  idCol = 1,
  dataCol = 2,
  buffSize = 1,
  transpose = FALSE,
  chunkEdge = 100
)
}
\arguments{
\item{dataFile}{file name of the text data}

\item{h5File}{file name of the result}

\item{headerRow}{row containing the header}

\item{idCol}{column with the ids}

\item{dataCol}{start of the data columns in the text data}

\item{buffSize}{buffer size}

\item{transpose}{boolean to transpose the data}

\item{chunkEdge}{chunk size for the hdf5 file}
}
\value{
Boolean stating if the result file was created succesfully
}
\description{
hdf5convert a function to convert textual data to hdf5 format supported by FastReg()
}
