% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FastReg2}
\alias{FastReg2}
\title{FastReg a function to perform fast simple linear or logistic regression}
\usage{
FastReg2(config.file, ...)
}
\arguments{
\item{config.file}{an optional character denoting configuration filename (default=NULL)}

\item{...}{additional parameters not specified in config.file}
}
\value{
numeric denoting elapsed.time
}
\description{
FastReg a function to perform fast simple linear or logistic regression
}
