% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_dataset.R
\name{generate_values}
\alias{generate_values}
\title{generate_values function to generate values for the test dataset}
\usage{
generate_values(num.ind, chunk.indices, poi.type, maf, miss.rate)
}
\arguments{
\item{num.ind}{an integer denoting number of individuals}

\item{poi.type}{character must be either "genotype" or "dosage"}

\item{maf}{MAF threshold}

\item{miss.rate}{rate for missing values}
}
\value{
a list consisting of dataset size
}
\description{
generate_values function to generate values for the test dataset
}
