% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_dataset.R
\name{generate_values2}
\alias{generate_values2}
\title{generate_values2 function to generate values for the test dataset}
\usage{
generate_values2(num.ind, poi.indices, poi.type, poi.chunk.size)
}
\arguments{
\item{num.ind}{an integer denoting number of individuals}

\item{poi.indices}{vector of POI ids}

\item{poi.type}{character must be either "genotype" or "dosage"}

\item{poi.chunk.size}{chunk size for the hdf5 dataset}
}
\value{
a list consisting of dataset size
}
\description{
generate_values2 function to generate values for the test dataset
}
