% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform-R-GLM-Match.R
\name{perform.GLM.test}
\alias{perform.GLM.test}
\title{perform.GLM.test function to validate FastReg results}
\usage{
perform.GLM.test(
  num.poi = 10,
  prefix = "testdata_500_by_500",
  input.dir = ".",
  output.dir = ".",
  transpose.h5 = FALSE,
  na.values = NA,
  mc.cores = 1
)
}
\arguments{
\item{num.poi}{an integer denoting number of POI to be evaluated}

\item{prefix}{a character string denoting filename prefix (same as one used by simulate_test_dataset function)}

\item{input.dir}{a character string denoting folder location input files}

\item{output.dir}{a character string denoting folder location of output files;}

\item{transpose.h5}{a logical denoting whether to transpose POI matrix}

\item{na.values}{vector of POI values that should treated as missing}

\item{mc.cores}{an integer denoting number of cores utilzed during parallel processing}
}
\value{
invisible TRUE when successful (POI level output for logistic and linear regression are saved in separate tsv files)
}
\description{
perform.GLM.test function to validate FastReg results
}
