% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{testdata_500_by_1k.bin.pheno}
\alias{testdata_500_by_1k.bin.pheno}
\title{testdata_500_by_1k.bin.pheno
Randomly generated binary phenotype data used for running vignettes
A compressed text file with 500 individuals x 1 binary response}
\format{
A data frame with 500 rows and 2 columns:
\describe{
  \item{ind}{\code{character} A unique alphanumeric identifier for each individual. Example: "IND123".}
  \item{bin.resp}{\code{integer} Binary phenotype data where 1 indicates the presence of a trait and 0 indicates its absence.}
}
}
\description{
testdata_500_by_1k.bin.pheno
Randomly generated binary phenotype data used for running vignettes
A compressed text file with 500 individuals x 1 binary response
}
