% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{testdata_500_by_1k.covar}
\alias{testdata_500_by_1k.covar}
\title{testdata_500_by_1k.covar
Randomly generated covariate data used for running vignettes
A compressed text file with 500 individuals x 3 covariates (age, sex, treatment)}
\format{
A data frame with 500 rows and 2 columns:
\describe{
  \item{ind}{\code{character} A unique alphanumeric identifier for each individual. Example: "IND123".}
  \item{age}{\code{integer} The age of the individual.}
  \item{sex}{\code{character} The sex of the individual - either M or F.}
  \item{treatment}{\code{character} The treatment that was administered to the individual. Test or Placebo.}
}
}
\description{
testdata_500_by_1k.covar
Randomly generated covariate data used for running vignettes
A compressed text file with 500 individuals x 3 covariates (age, sex, treatment)
}
