% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{testdata_500_by_1k.num.pheno}
\alias{testdata_500_by_1k.num.pheno}
\title{testdata_500_by_1k.num.pheno
Randomly generated numerical phenotye data used for running vignettes
A compressed text file with 500 individuals x 1 numerical response}
\format{
A data frame with 500 rows and 2 columns:
\describe{
  \item{ind}{\code{character} A unique alphanumeric identifier for each individual. Example: "IND123".}
  \item{num.resp}{\code{integer} Numerical phenotype data.}
}
}
\description{
testdata_500_by_1k.num.pheno
Randomly generated numerical phenotye data used for running vignettes
A compressed text file with 500 individuals x 1 numerical response
}
