% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{testdata_500_by_1k.poi}
\alias{testdata_500_by_1k.poi}
\title{testdata_500_by_1k.poi
Randomly generated binary POI data used for running vignettes
A compressed text file with 500 individuals x 1000 binary responses}
\format{
A data frame with 500 rows and 1000 columns:
\describe{
  \item{ind}{\code{character} A unique alphanumeric identifier for each individual. Example: "IND123".}
  \item{rs0001}{\code{integer} Binary response for the 1st poi. May contain no response at all.}
  \item{rs0002}{\code{integer} Binary response for the 2nd poi. May contain no response at all.}
  \item{rs####}{\code{integer} Binary response for the nth poi. May contain no response at all.}
}
}
\description{
testdata_500_by_1k.poi
Randomly generated binary POI data used for running vignettes
A compressed text file with 500 individuals x 1000 binary responses
}
