% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TextToH5}
\alias{TextToH5}
\title{TextToH5 a function to convert textual data to hdf5 format supported by FastReg()}
\usage{
TextToH5(
  dataFile,
  h5Dir,
  headerRow = 1,
  idCol = 1,
  dataCol = 2,
  buffSize = 1,
  transpose = FALSE,
  chunkEdge = 100,
  vcf = FALSE,
  delimiter = " \\t",
  gz = FALSE,
  poiPerFile = -1,
  singleFile = FALSE,
  serverThreads = -1,
  serverMem = -1
)
}
\arguments{
\item{dataFile}{Path of the input text file.}

\item{h5Dir}{Path of a new directory to write generated hdf5 files.}

\item{headerRow}{Default: 1. Row number of header line.}

\item{idCol}{Default 1. Column number that contains identifiers.}

\item{dataCol}{Default: 2. Column number that contains first data entry.}

\item{buffSize}{Default: 1.0. Size in Gb of memory available for data conversion.}

\item{transpose}{Default: FALSE. Boolean value that indicates whether text data should be transposed. Set TRUE when columns represent individuals and rows represent POIs.}

\item{chunkEdge}{Default: 100. Specify size of data chunks to be applied to output hdf5 files. Must be less than or equal to values of poiPerFile and greater than zero.}

\item{vcf}{Default: FALSE. Indicate whether input file is in vcf format. If TRUE, headerRow, idCol, dataCol and transpose values are ignored.}

\item{delimiter}{Specify a string of all delimiters used in the text file to separate columns. Default will split entries by space or tab.}

\item{gz}{Default: FALSE. Indicate whether input text file is gzip-compressed.}

\item{poiPerFile}{Default: -1. Indicate the number of POIs to write to each output hdf5 file. A value of -1 indicates the count should be calculated based on available system resources.}

\item{singleFile}{Default: FALSE. Indicate whether a single hdf5 file should be produced rather than a series.}

\item{serverThreads}{Default: -1. Indicate the total number of CPU threads available for FastReg - utilized to determine the optimal number of output files. A value of -1 indicates the count should be detected automatically. Set this parameter when running FastReg on a shared system and specify resources allocated for your own use only.}

\item{serverMem}{Default: -1.0. Indicate the total memory in Gb available for FastReg - utilized to determine the optimal number of output files. A value of -1.0 indicates the total should be detected automatically. Set this parameter when running FastReg on a shared system and specify resources allocated for your own use only.}
}
\value{
0 = success, 1 = failure
}
\description{
TextToH5 a function to convert textual data to hdf5 format supported by FastReg()
}
