% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_dataset.R
\name{write.h5}
\alias{write.h5}
\title{write.h5 function to write the hdf5 dataset}
\usage{
write.h5(
  data.num.chunks,
  prefix,
  verbose,
  ind.id,
  poi.id,
  poi.data.dir,
  poi.compression.level,
  poi.chunk.size,
  poi.type,
  num.ind,
  num.poi
)
}
\arguments{
\item{data.num.chunks}{number of files to split the data into}

\item{prefix}{prefix for all files created}

\item{verbose}{logical to control display of progress messages (default=TRUE)}

\item{ind.id}{array of unique individual ids}

\item{poi.id}{array of unique POI ids}

\item{poi.data.dir}{output directory for hdf5 dataset}

\item{poi.compression.level}{gzip compression level for hdf5}

\item{poi.chunk.size}{chunk size for the hdf5 dataset}

\item{poi.type}{character must be either "genotype" or "dosage"}

\item{num.ind}{an integer denoting number of individuals}

\item{num.poi}{an integer denoting number of predictors of interest (SNP calls)}
}
\value{
none
}
\description{
write.h5 function to write the hdf5 dataset
}
