% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastR_hdf5convert.R
\name{fastR_hdf5convert}
\alias{fastR_hdf5convert}
\title{hdf5convert a function to convert textual data to hdf5 format supported by FastReg()}
\usage{
fastR_hdf5convert(
  dataFile,
  h5Dir,
  headerRow = 1,
  idCol = 1,
  dataCol = 2,
  buffSize = 1,
  transpose = FALSE,
  chunkEdge = 100,
  vcf = FALSE,
  delimiter = " \\t",
  gz = FALSE,
  poiPerFile = -1,
  singleFile = FALSE,
  serverThreads = -1,
  serverMem = -1
)
}
\arguments{
\item{dataFile}{file name of the text data}

\item{h5Dir}{file dir to write hdf5 files to}

\item{headerRow}{row containing the header}

\item{idCol}{column with the ids}

\item{dataCol}{start of the data columns in the text data}

\item{buffSize}{buffer size}

\item{transpose}{boolean to transpose the data}

\item{chunkEdge}{chunk size for the hdf5 file}

\item{vcf}{TRUE or FALSE}

\item{delimiter}{delimiter used in the dataFile}

\item{gz}{bool to gzip the h5 dataset or not}

\item{poiPerFile}{=-1}

\item{singleFile}{=FALSE}

\item{serverThreads}{=-1}

\item{serverMem}{=-1.0}
}
\value{
Boolean stating if the result file was created succesfully
}
\description{
hdf5convert a function to convert textual data to hdf5 format supported by FastReg()
}
