% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FastReg}
\alias{FastReg}
\title{FastReg a function to perform fast simple linear or logistic regression}
\usage{
FastReg(
  phenotype = "bin.resp",
  regression.type = "logistic",
  Pvalue.dist = "t.dist",
  output.exclude.covar = TRUE,
  maf.threshold = 1e-13,
  hwe.threshold = 1e-13,
  no.intercept = FALSE,
  colinearity.rsq = 1,
  poi.block.size = 0,
  max.iter = 6,
  rel.conv.tolerance = 0.01,
  abs.conv.tolerance = 0.01,
  max.openmp.threads = 1,
  pheno.file = "testdata_1k_by_5k.bin.pheno.txt",
  pheno.rowname.cols = "ind",
  pheno.file.delim = "tab",
  covar.file = "testdata_1k_by_5k.covar.txt",
  covar.rowname.cols = "ind",
  covar.file.delim = "tab",
  POI.file.dir = "testdata/",
  POI.file.delim = "tab",
  POI.file.format = "h5",
  POI.type = "dosage",
  POI.effect.type = "additive",
  covariates = c("age", "sex", "eth", "treatment", "severity"),
  covariate.type = c("numeric", "categorical", "categorical", "categorical",
    "categorical"),
  covariate.standardize = c(TRUE, FALSE, FALSE, FALSE, FALSE),
  covariate.levels = c("", "F,M", "afr,asi,eur", "Placebo,Test",
    "Very Low,Low,Moderate,High,Very High,Extreme"),
  covariate.ref.level = c("", "F", "eur", "Placebo", "Very Low"),
  POI.covar.interactions = c(""),
  split.by = c(""),
  output.dir = "test",
  compress.results = FALSE,
  max.blas.threads = -1,
  max.workers = 0
)
}
\arguments{
\item{phenotype}{bin.resp | num.resp. Default: bin.resp}

\item{regression.type}{logistic | linear. Default: logistic}

\item{Pvalue.dist}{t.dist | norm.dist. Default: t.dist}

\item{output.exclude.covar}{TRUE | FALSE. Default: TRUE. Excludes covariates in the results output}

\item{maf.threshold}{Default: 1e-13. Minor Allele frequency threshold. Acceptable values: 0-0.5}

\item{hwe.threshold}{Default: 1e-13. Hardy-Weinberg Equilibrium threshold. Acceptable values: 0-0.5}

\item{no.intercept}{TRUE | FALSE. Default: FALSE.}

\item{colinearity.rsq}{Default: 1.0. Filter covariates if above R-squared colinearity threshold. Acceptage values: 0.8-1.0}

\item{poi.block.size}{Default: 0. Overrides auto-calculated poi block size.}

\item{max.iter}{Default: 6. Number of logistic regression iterations. Doesn't apply when regression.type = linear.}

\item{rel.conv.tolerance}{Default: 0.01. Relative convergence threshold for POIs.}

\item{abs.conv.tolerance}{Default: 0.01. Absolute convergence threshold for POIs.}

\item{max.openmp.threads}{Default: 1. Overrides number of threads used. By default FastReg will use 1 openmp threads.}

\item{pheno.file}{relative path to phenotype file. Contains outcomes that will be modelled.}

\item{pheno.rowname.cols}{Default: ind. column to be treated as the subject identifier when matching across files. Can be multiple columns (comma separated).}

\item{pheno.file.delim}{tab|space|comma. Default: tab. delimiter used in phenotype file.}

\item{covar.file}{relative path to covariates file.}

\item{covar.rowname.cols}{Default: ind. column to be treated as the subject identifier when matching across files. Can be multiple columns (comma separated).}

\item{covar.file.delim}{tab|space|comma. Default: tab. delimiter used in covariate file.}

\item{POI.file.dir}{relative path to the directory with POI file(s).}

\item{POI.file.delim}{tab|space|comma. Default: tab. delimiter used in POI file if POI.file.format != h5}

\item{POI.file.format}{h5|txt. Default: h5. POI data file format.}

\item{POI.type}{genotype|dosage. Default: dosage.}

\item{POI.effect.type}{additive|dominant|recessive. Default: additive.}

\item{covariates}{Default: c("age","sex","eth","treatment","severity"). List of column names of the covariates file.}

\item{covariate.type}{Default: c("numeric","categorical","categorical","categorical","categorical"). List of column data type for each covariate listed in covariates.}

\item{covariate.standardize}{Default: c(TRUE,FALSE,FALSE,FALSE,FALSE). List of booleans representing which covariates are to be standardized.}

\item{covariate.levels}{Default: c("", "F,M","afr,asi,eur","Placebo,Test","Very Low,Low,Moderate,High,Very High,Extreme"). List of unique values for categorical columns. (comma separated for each unique value)}

\item{covariate.ref.level}{Default: c("","F","eur","Placebo","Very Low"). List of reference level for covariate value (comma separated)}

\item{POI.covar.interactions}{Default: c(""). List of poi-covariate interactions}

\item{split.by}{Default: c(""). List of covariates values to split by when stratifying the data.}

\item{output.dir}{Default: results. Relative path to directory for output. Directory will be created if it doesn't exist}

\item{compress.results}{TRUE|FALSE. Default: FALSE. Compress results inside the output directory.}

\item{max.blas.threads}{Default: -1. Set the number of BLAS threads that Eigen/Armadillo will use. The default value will use the system/BLAS library default.}

\item{max.workers}{Default: 0. Number of processes FastReg will spawn. By default FastReg will spawn min(# of poi files, # of available workers)}
}
\value{
numeric denoting elapsed.time
}
\description{
FastReg a function to perform fast simple linear or logistic regression
}
